reorganising storage module
