+++ src/benchmark/tatp_workload.h
@@ -29,7 +29,7 @@
 #include "catalog/column_schema.h"
 #include "catalog/table_schema.h"
 #include "sqlite3/sqlite3.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 namespace hustle::operators {
 class TATP {
