+++ src/main.cc
@@ -9,7 +9,7 @@
 #include "catalog/column_schema.h"
 #include "catalog/table_schema.h"
 #include "sqlite3/sqlite3.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 // The function to be executed by all threads
 void *readQuery(void *db) {
