+++ src/operators/aggregate/aggregate.cc
@@ -23,7 +23,7 @@
 #include <iostream>
 #include <utility>
 
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 #define AGGREGATE_OUTPUT_TABLE "aggregate"
