+++ src/operators/fused/select_build_hash.cc
@@ -26,7 +26,7 @@
 #include <utility>
 
 #include "storage/table.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 #include "utils/bit_utils.h"
 
 namespace hustle {
