+++ src/operators/join/join.cc
@@ -23,7 +23,7 @@
 #include <iostream>
 #include <utility>
 
-#include "storage/util.h"
+#include "storage/utils/util.h"
 #include "utils/bloom_filter.h"
 
 namespace hustle::operators {
