+++ src/operators/tests/select_build_hash_test.cc
@@ -27,7 +27,7 @@
 #include "gtest/gtest.h"
 #include "scheduler/scheduler.h"
 #include "storage/block.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 #define BLOCK_SIZE 108
