+++ src/storage/block.cc
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "absl/strings/numbers.h"
-#include "util.h"
+#include "storage/utils/util.h"
 
 #define ESTIMATED_STR_LEN 30
