+++ src/storage/tests/block_test.cc
@@ -23,7 +23,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "storage/table.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 #define BLOCK_SIZE \
   1024  // Force the block size to 1 KB for the sake of this test
