+++ src/storage/tests/table_test.cc
@@ -26,7 +26,7 @@
 #include "api/hustle_db.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 #define BLOCK_SIZE \
   1024  // Force the block size to 1 KB for the sake of this test
