+++ src/utils/arrow_compute_wrappers.cc
@@ -21,7 +21,7 @@
 
 #include <thread>
 
-#include "storage/util.h"
+#include "storage/utils/util.h"
 
 namespace hustle {
